requirements.txt
##############################
# Core API Framework
##############################
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.0.0

##############################
# LangChain / LangGraph Stack
##############################
langgraph>=1.0.0,<2.0.0
langchain>=1.0.0,<2.0.0
langchain-core>=1.0.0,<2.0.0
langchain-text-splitters>=0.3.0,<1.0.0

##############################
# Google Gemini (New SDK)
##############################
google-genai>=1.0.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-cloud-storage>=2.18.0

##############################
# Pinecone (Vector DB - New Package)
##############################
pinecone>=5.0.0

##############################
# CLIP + Vision Embeddings
##############################
torch>=2.3.0,<3.0.0
torchvision>=0.18.0,<1.0.0
open-clip-torch>=3.2.0
ftfy>=6.2.0
regex>=2024.0.0
Pillow>=11.0.0
numpy>=1.26.0,<3.0.0

##############################
# Utilities / Networking
##############################
requests>=2.32.0,<3.0.0
httpx>=0.27.0,<1.0.0
python-dotenv>=1.0.0

##############################
# Document Processing (Attachments)
##############################
PyMuPDF>=1.24.0          # PDF text extraction (primary)
python-docx>=1.1.0       # Word document extraction
openpyxl>=3.1.0          # Excel spreadsheet extraction
lxml>=5.0.0              # XML/HTML parsing (dependency)

##############################
# Dev Tools / Debugging
##############################
graphviz>=0.20.0
IPython>=8.0.0
python-dateutil>=2.9.0

##############################
# Additional utilities
##############################
colorlog>=6.0.0
diskcache>=5.0.0
tqdm>=4.60.0
